+++ product-catalogue-frontend/src/app/services/product.service.ts
@@ -11,62 +11,67 @@ import { AuthenticationService } from './authentication.service';
 })
 export class ProductService {
 
-  constructor(private httpClient: HttpClient, private authService:AuthenticationService) {
+  constructor(private httpClient: HttpClient, private authService: AuthenticationService) {
 
   }
 
   public getProductsByCategory(uri: string): Observable<Product> {
-    return this.httpClient.get<Product>(uri);
+    return this.httpClient.get<Product>(uri, { headers: new HttpHeaders({ 'Authorization': "Bearer " + this.authService.getJwtAccessToken() }) });
   }
 
   getAllProducts(uri: string): Observable<Product> {
-    return this.httpClient.get<Product>(uri,{headers:new HttpHeaders({'Authorization':"Bearer "+this.authService.getJwtAccessToken()})});
+    if (this.authService.isJwtExpired()) {
+      // traiter le cas de manipulation du refresh-token
+      throw new Error("JWT Access Token est expiré!");
+    }
+    let authorizationHeader = new HttpHeaders({ 'Authorization': "Bearer " + this.authService.getJwtAccessToken() });
+    return this.httpClient.get<Product>(uri, { headers: authorizationHeader });
   }
 
   getProducts(pageNbr: number, size: number): Observable<Product> {
-    return this.httpClient.get<Product>(GlobalService.HOST + "/products?page=" + pageNbr + "&size=" + size,{headers:new HttpHeaders({'Authorization':"Bearer "+this.authService.getJwtAccessToken()})});
+    return this.httpClient.get<Product>(GlobalService.HOST + "/products?page=" + pageNbr + "&size=" + size, { headers: new HttpHeaders({ 'Authorization': "Bearer " + this.authService.getJwtAccessToken() }) });
   }
 
   searchProductsByKeyword(keyword: string, pageNbr: number, size: number): Observable<Product> {
-    return this.httpClient.get<Product>(GlobalService.HOST + "/products/search/filterByDesignationPage?key=" + keyword + "&page=" + pageNbr + "&size=" + size,{headers:new HttpHeaders({'Authorization':"Bearer "+this.authService.getJwtAccessToken()})});
+    return this.httpClient.get<Product>(GlobalService.HOST + "/products/search/filterByDesignationPage?key=" + keyword + "&page=" + pageNbr + "&size=" + size, { headers: new HttpHeaders({ 'Authorization': "Bearer " + this.authService.getJwtAccessToken() }) });
   }
 
   deleteProduct(URL: string) {
     // Noter bien que la mtd DELETE ne retourne rien
-    return this.httpClient.delete(URL);
+    return this.httpClient.delete(URL, { headers: new HttpHeaders({ 'Authorization': "Bearer " + this.authService.getJwtAccessToken() }) });
   }
 
   updateProductAssociation(URIOfProductToBindTo: string, URIOfCategoryToBind: string) {
-    return this.httpClient.put(URIOfProductToBindTo, URIOfCategoryToBind, { headers: new HttpHeaders({ 'Content-Type': 'text/uri-list','Authorization':"Bearer "+this.authService.getJwtAccessToken() }) });
+    return this.httpClient.put(URIOfProductToBindTo, URIOfCategoryToBind, { headers: new HttpHeaders({ 'Content-Type': 'text/uri-list', 'Authorization': "Bearer " + this.authService.getJwtAccessToken() }) });
 
   }
 
   updateProduct(URL: string, dataOfEditedProduct: any) {
-    return this.httpClient.put(URL, dataOfEditedProduct,{headers:new HttpHeaders({'Authorization':"Bearer "+this.authService.getJwtAccessToken()})});
+    return this.httpClient.put(URL, dataOfEditedProduct, { headers: new HttpHeaders({ 'Authorization': "Bearer " + this.authService.getJwtAccessToken() }) });
 
   }
 
   patchProduct(URL: string, dataOfEditedProduct: any) {
-    return this.httpClient.patch<Product>(URL, dataOfEditedProduct,{headers:new HttpHeaders({'Authorization':"Bearer "+this.authService.getJwtAccessToken()})});
+    return this.httpClient.patch<Product>(URL, dataOfEditedProduct, { headers: new HttpHeaders({ 'Authorization': "Bearer " + this.authService.getJwtAccessToken() }) });
 
   }
 
   saveProduct(URL: string, data: Product): Observable<Product> {
     // Noter bien que la mtd POST RETROUNE l objet enregistre format JSON, avec ses propres _links
-    return this.httpClient.post<Product>(URL, data,{headers:new HttpHeaders({'Authorization':"Bearer "+this.authService.getJwtAccessToken()})});
+    return this.httpClient.post<Product>(URL, data, { headers: new HttpHeaders({ 'Authorization': "Bearer " + this.authService.getJwtAccessToken() }) });
   }
 
   getProduct(URL: string): Observable<Product> {
-    return this.httpClient.get<Product>(URL,{headers:new HttpHeaders({'Authorization':"Bearer "+this.authService.getJwtAccessToken()})});
+    return this.httpClient.get<Product>(URL, { headers: new HttpHeaders({ 'Authorization': "Bearer " + this.authService.getJwtAccessToken() }) });
   }
 
   getProductPhoto(URL: string) {
-    return this.httpClient.get(URL,{headers:new HttpHeaders({'Authorization':"Bearer "+this.authService.getJwtAccessToken()})});
+    return this.httpClient.get(URL, { headers: new HttpHeaders({ 'Authorization': "Bearer " + this.authService.getJwtAccessToken() }) });
     //pas besoin ,{ headers: new HttpHeaders({ 'Content-Type': 'image/png' }) }
   }
 
   getCategoryOfProduct(URICategory: string): Observable<Category> {
-    return this.httpClient.get<Category>(URICategory,{headers:new HttpHeaders({'Authorization':"Bearer "+this.authService.getJwtAccessToken()})});
+    return this.httpClient.get<Category>(URICategory, { headers: new HttpHeaders({ 'Authorization': "Bearer " + this.authService.getJwtAccessToken() }) });
   }
 
   /*
